CheckoutTwo Methods Updated
Integer return is now a double return.

+++ src/test/java/com/feudaloverlords/swaglabs/pom/CheckoutTwoPage.java
@@ -29,18 +29,18 @@ public class CheckoutTwoPage extends Page {
         super(driver, "https://www.saucedemo.com/checkout-step-two.html");
     }
 
-    public Integer getItemTotal (WebDriver webDriver){
-        Integer itemTotal = Integer.valueOf(driver.findElement(new By.ByClassName("summary_subtotal_label")).getText());
+    public double getItemTotal (WebDriver webDriver){
+        double itemTotal = Double.valueOf(driver.findElement(new By.ByClassName("summary_subtotal_label")).getText());
         return itemTotal;
     }
 
-    public Integer getTotal (WebDriver webDriver){
-        Integer total = Integer.valueOf(driver.findElement(new By.ByClassName("summary_total_label")).getText());
+    public double getTotal (WebDriver webDriver){
+        double total = Double.valueOf(driver.findElement(new By.ByClassName("summary_total_label")).getText());
         return total;
     }
 
-    public Integer getTax (WebDriver webDriver){
-        Integer tax = Integer.valueOf(driver.findElement(new By.ByClassName("summary_tax_label")).getText());
+    public double getTax (WebDriver webDriver){
+        double tax = Double.valueOf(driver.findElement(new By.ByClassName("summary_tax_label")).getText());
         return tax;
     }
